替换为Eclipse Collections

+++ src/main/java/com/QYun/AssetReader4J/SevenZip/LZMA/Encoder.java
@@ -34,13 +34,11 @@ public class Encoder {
     Optimal[] _optimum = new Optimal[kNumOpts];
     com.QYun.AssetReader4J.SevenZip.LZ.BinTree _matchFinder = null;
 
-    ;
     com.QYun.AssetReader4J.SevenZip.RangeCoder.Encoder _rangeEncoder = new com.QYun.AssetReader4J.SevenZip.RangeCoder.Encoder();
     short[] _isMatch = new short[Base.kNumStates << Base.kNumPosStatesBitsMax];
     short[] _isRep = new short[Base.kNumStates];
     short[] _isRepG0 = new short[Base.kNumStates];
 
-    ;
     short[] _isRepG1 = new short[Base.kNumStates];
     short[] _isRepG2 = new short[Base.kNumStates];
     short[] _isRep0Long = new short[Base.kNumStates << Base.kNumPosStatesBitsMax];
@@ -173,7 +171,7 @@ public class Encoder {
         if (_numDistancePairs > 0) {
             lenRes = _matchDistances[_numDistancePairs - 2];
             if (lenRes == _numFastBytes)
-                lenRes += _matchFinder.GetMatchLen((int) lenRes - 1, _matchDistances[_numDistancePairs - 1],
+                lenRes += _matchFinder.GetMatchLen(lenRes - 1, _matchDistances[_numDistancePairs - 1],
                         Base.kMatchMaxLen - lenRes);
         }
         _additionalOffset++;
@@ -760,10 +758,10 @@ public class Encoder {
             int complexState = (_state << Base.kNumPosStatesBitsMax) + posState;
             if (len == 1 && pos == -1) {
                 _rangeEncoder.Encode(_isMatch, complexState, 0);
-                byte curByte = _matchFinder.GetIndexByte((int) (0 - _additionalOffset));
+                byte curByte = _matchFinder.GetIndexByte(0 - _additionalOffset);
                 LiteralEncoder.Encoder2 subCoder = _literalEncoder.GetSubCoder((int) nowPos64, _previousByte);
                 if (!Base.StateIsCharState(_state)) {
-                    byte matchByte = _matchFinder.GetIndexByte((int) (0 - _repDistances[0] - 1 - _additionalOffset));
+                    byte matchByte = _matchFinder.GetIndexByte(0 - _repDistances[0] - 1 - _additionalOffset);
                     subCoder.EncodeMatched(_rangeEncoder, matchByte, curByte);
                 } else
                     subCoder.Encode(_rangeEncoder, curByte);
@@ -810,7 +808,7 @@ public class Encoder {
                     _posSlotEncoder[lenToPosState].Encode(_rangeEncoder, posSlot);
 
                     if (posSlot >= Base.kStartPosModelIndex) {
-                        int footerBits = (int) ((posSlot >> 1) - 1);
+                        int footerBits = (posSlot >> 1) - 1;
                         int baseVal = ((2 | (posSlot & 1)) << footerBits);
                         int posReduced = pos - baseVal;
 
@@ -927,7 +925,7 @@ public class Encoder {
     void FillDistancesPrices() {
         for (int i = Base.kStartPosModelIndex; i < Base.kNumFullDistances; i++) {
             int posSlot = GetPosSlot(i);
-            int footerBits = (int) ((posSlot >> 1) - 1);
+            int footerBits = (posSlot >> 1) - 1;
             int baseVal = ((2 | (posSlot & 1)) << footerBits);
             tempPrices[i] = BitTreeEncoder.ReverseGetPrice(_posEncoders,
                     baseVal - posSlot - 1, footerBits, i - baseVal);
@@ -1218,7 +1216,6 @@ public class Encoder {
 
         public void MakeAsShortRep() {
             BackPrev = 0;
-            ;
             Prev1IsChar = false;
         }
